requirements.txt
fastapi==0.112.2
uvicorn==0.30.6
google==3.0.0
google-api-core==2.8.1
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-cloud-bigquery==3.18.0
google-cloud-bigquery-storage==2.13.1
google-cloud-core==2.3.0
google-cloud-storage==2.3.0
google-crc32c==1.3.0
google-pasta==0.2.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.2
poetry==1.8.3
python-dotenv==1.0.0
python-multipart==0.0.6
pyautogen==0.2.34
pytz==2021.3
requests==2.26.0
bs4==0.0.1
image==1.5.33
matplotlib==3.9.2
matplotlib-inline==0.1.6
seaborn==0.12.2
numpy==1.26.4
pandas==2.2.2
scipy==1.14.1
squarify==0.4.3
unstructured==0.10.25
pyowm==3.3.0
selenium==4.23.1
openai==1.42.0
tiktoken==0.5.2
firecrawl-py==0.0.20
twine==5.1.1
twilio==9.2.3
tavily-python==0.4.0
fonttools==4.53.1
markdown2==2.5.0
fpdf2==2.7.9
pyppeteer==2.0.0
